pyproject.toml
[project]
name = "caramelo-papelaria"
version = "1.0.0"
description = "Static link page generator for Caramelo Papelaria"
readme = "readme.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0",
    "pytest-cov>=7.0",
]

[project.scripts]
generate-site = "generate_site:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = [
    ".venv/*",
    "tests/*",
    "docs/*",
]
